+++ src/pyxspress/util/__init__.py
@@ -1,2 +1,2 @@
 from .loggable import Loggable
-from .util import setup_basic_logging
+from .util import get_module_logger, setup_basic_logging