+++ test_files/config_8Ch/stControlServer.sh
@@ -2,6 +2,6 @@
 SCRIPT_DIR="$( cd "$( dirname "$0" )" && pwd )"
 
 export LD_LIBRARY_PATH=/odin/xspress/lib
-export XSP3READOUT=passive
+export XSP3READOUT=frames
 
 /odin/prefix/bin/xspressControl -j $SCRIPT_DIR/xspress.json